packages/daw-mcp/pyproject.toml
[tool.poetry]
name = "daw-mcp"
version = "0.1.0"
description = "DAW Custom MCP Servers - Git and Graph Memory"
authors = ["RADit Team <team@radit.dev>"]
readme = "README.md"
packages = [{ include = "daw_mcp", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
mcp = "^1.9.0"
neo4j = "^6.0.3"
gitpython = "^3.1.43"
pydantic = "^2.12.5"
anyio = "^4.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
mypy = "^1.19.1"
ruff = "^0.14.10"

[tool.poetry.scripts]
daw-git-mcp = "daw_mcp.git_mcp.server:main"
daw-graph-memory = "daw_mcp.graph_memory.server:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --strict-markers"
markers = [
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["git.*", "neo4j.*"]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP"]
ignore = [
    "E501",  # Line too long
    "UP017",  # datetime.UTC requires Python 3.11+
]
